Open git dialog in rad push action not in rad sync.Add rad pull functionality
