+++ src/main/java/network/radicle/jetbrains/radiclejetbrainsplugin/RadicleSyncAction.java
@@ -1,37 +1,25 @@
 package network.radicle.jetbrains.radiclejetbrainsplugin;
 
 import com.google.common.base.Strings;
-import com.intellij.dvcs.DvcsUtil;
-import com.intellij.dvcs.push.PushInfo;
-import com.intellij.dvcs.push.ui.VcsPushDialog;
-import com.intellij.dvcs.repo.Repository;
-import com.intellij.dvcs.repo.VcsRepositoryManager;
 import com.intellij.notification.Notification;
 import com.intellij.notification.NotificationAction;
 import com.intellij.notification.NotificationGroupManager;
 import com.intellij.notification.NotificationType;
 import com.intellij.openapi.actionSystem.AnAction;
 import com.intellij.openapi.actionSystem.AnActionEvent;
-import com.intellij.openapi.actionSystem.CommonDataKeys;
 import com.intellij.openapi.application.ApplicationManager;
 import com.intellij.openapi.options.ShowSettingsUtil;
 import com.intellij.openapi.project.Project;
-import com.intellij.openapi.vfs.VirtualFile;
-import com.intellij.util.containers.ContainerUtil;
 import git4idea.repo.GitRepository;
 import git4idea.repo.GitRepositoryManager;
 import network.radicle.jetbrains.radiclejetbrainsplugin.config.RadicleSettingsHandler;
 import network.radicle.jetbrains.radiclejetbrainsplugin.config.RadicleSettingsView;
 import network.radicle.jetbrains.radiclejetbrainsplugin.services.RadicleApplicationService;
-import network.radicle.jetbrains.radiclejetbrainsplugin.services.RadicleProjectService;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.util.Collection;
-import java.util.Collections;
-import java.util.HashSet;
 import java.util.List;
 import java.util.function.Supplier;
 
@@ -60,15 +48,14 @@ public class RadicleSyncAction extends AnAction {
             logger.warn("no git repos found!");
             return;
         }
-        openGitPushDialog(project, e);
-        var rps = project.getService(RadicleProjectService.class);
-        rps.forceRadSync = true;
+        ApplicationManager.getApplication().executeOnPooledThread(() -> {
+            sync(repos, project);
+        });
     }
 
-    public static void sync(List<PushInfo> pushDetails, Project project) {
+    public static void sync(List<GitRepository> repos, Project project) {
         var rad = ApplicationManager.getApplication().getService(RadicleApplicationService.class);
-        for (var pi : pushDetails) {
-            var repo = (GitRepository) pi.getRepository();
+        for (var repo : repos) {
             // TODO: most probably there is a better way to understand if this repo is rad enabled
             var output = rad.sync(repo);
             var success = output.checkSuccess(com.intellij.openapi.diagnostic.Logger.getInstance(RadicleApplicationService.class));
@@ -81,30 +68,6 @@ public class RadicleSyncAction extends AnAction {
             showNotification(project, "", "Synced project with radicle seed node", NotificationType.INFORMATION, null);
         }
         // TODO: could be useful: PersistentDefaultAccountHolder and GithubProjectDefaultAccountHolder
-    }
-
-    private void openGitPushDialog(Project project , @NotNull AnActionEvent e) {
-        VcsRepositoryManager manager = VcsRepositoryManager.getInstance(project);
-        Collection<Repository> repositories = e.getData(CommonDataKeys.EDITOR) != null
-                ? ContainerUtil.emptyList()
-                : collectRepositories(manager, e.getData(CommonDataKeys.VIRTUAL_FILE_ARRAY));
-        VirtualFile selectedFile = DvcsUtil.getSelectedFile(project);
-        new VcsPushDialog(project, DvcsUtil.sortRepositories(repositories),
-                selectedFile != null ? manager.getRepositoryForFileQuick(selectedFile) : null).show();
-    }
-
-    @NotNull
-    private static Collection<Repository> collectRepositories(@NotNull VcsRepositoryManager vcsRepositoryManager,
-                                                              VirtualFile @Nullable [] files) {
-        if (files == null) return Collections.emptyList();
-        Collection<Repository> repositories = new HashSet<>();
-        for (VirtualFile file : files) {
-            Repository repo = vcsRepositoryManager.getRepositoryForFileQuick(file);
-            if (repo != null) {
-                repositories.add(repo);
-            }
-        }
-        return repositories;
     }
 
     public static class ConfigureRadCliNotificationAction extends NotificationAction {
